requirements.txt
# ── Core Framework ──
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# ── Authentication ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ── Frontend ──
streamlit==1.40.2
pandas==2.2.3
requests==2.32.3
plotly==5.24.1

# ── Dev / Utilities ──
python-dotenv==1.0.1